Fix mpiib parse result file

+++ src/parse-results/mpiib.py
@@ -18,7 +18,7 @@ def main():
         res = re.findall("\d+\s+(\d+(\.\d+)?)", line, re.MULTILINE)
         if (res):
             print("Expected max network bandwidth: "
-                  + str(int(res[len(res) - 1][0] if type(res[len(res) - 1]) is tuple else res[len(res) - 1]) * 1000000)
+                  + str(float(res[len(res) - 1][0] if type(res[len(res) - 1]) is tuple else res[len(res) - 1]) * 1000000)
                   + " MB/s in job with ID "
                   + resultFileName.replace("slurm-", "").replace(".out", ""))
             if ('--dndoof' not in sys.argv):
